requirements.txt
asgiref==3.7.2
Django==4.2.7
django-taggit==5.0.1
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
